+++ railties/lib/rails/application.rb
@@ -322,7 +322,8 @@ module Rails
         if File.exist?(yaml)
           require "erb"
           all_secrets = YAML.load(ERB.new(IO.read(yaml)).result) || {}
-          env_secrets = all_secrets[Rails.env]
+          env_secrets = all_secrets['all'] || {}
+          env_secrets.merge!(all_secrets[Rails.env]) if all_secrets[Rails.env]
           secrets.merge!(env_secrets.symbolize_keys) if env_secrets
         end